api/requirements.txt
boto3
django
djangorestframework
django-filter
mysqlclient
django-environ
pymysql
django-rest-swagger
drf-yasg
django-redis-sessions
uwsgi
django-cors-headers
